genai/pyproject.toml
[tool.poetry]
name = "quiz-genai"
version = "0.1.0"
description = ""
authors = ["Protons (proton98)"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pydantic = "^2.10.3"
weaviate-client = "^4.9.6"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
langchain = "^0.3.9"
langchain-openai = "^0.2.11"
celery = "^5.4.0"
pymongo = {extras = ["srv"], version = "^4.10.1"}
prometheus-client = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
